chore: refactor data objects and paths
