+++ docs/.vitepress/theme/components/Helpers/AttributionLink.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import CardSingle from '@theme/components/CardLists/CardSingle.vue';
-import type { Person } from '../../../../data';
+import type { Person } from '@theme/data/people';
 
 defineProps<{
   attribution: Person;
