+++ docs/.vitepress/theme/components/Helpers/AttributionLinks.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-  import type { PeopleKeys, Person } from '../../../../data';
+  import type { Person } from '@theme/data/people';
   import AttributionLink from '@theme/components/Helpers/AttributionLink.vue';
 
   defineProps<{
