+++ docs/.vitepress/theme/components/WCUsedByProjects/ProjectsUsingWCCardList.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { projectsUsingWCCardData } from '../../../../data';
+import { projectsUsingWCCardData } from '@theme/data/home';
 import CardList from '@theme/components/CardLists/CardList.vue';
 import ProjectsUsingWCCard from './ProjectsUsingWCCard.vue';
 </script>
