+++ docs/.vitepress/theme/data/home.ts
@@ -1,47 +1,3 @@
-export const guideLinks = [
-  { text: 'Introduction', link: '/guides/introduction' },
-  { text: 'Quickstart', link: '/guides/quickstart' },
-  { text: 'Working with the File System', link: '/guides/working-with-the-file-system' },
-  { text: 'Running Processes', link: '/guides/running-processes' },
-  { text: 'Configuring Headers', link: '/guides/configuring-headers' },
-  { text: 'Troubleshooting', link: '/guides/troubleshooting'},
-  { text: 'Browser Support', link: '/guides/browser-support' },
-  { text: 'Browser Configuration', link: '/guides/browser-config' },
-  // { text: 'Community Inspirations', link: '/guides/community-inspirations' },
-];
-
-export const tutorialLinks = [
-  { text: '1. Build your first WebContainer app', link: '/tutorial/1-build-your-first-webcontainer-app' },
-  { text: '2. Setting up WebContainers', link: '/tutorial/2-setting-up-webcontainers' },
-  { text: '3. Installing dependencies', link: '/tutorial/3-installing-dependencies' },
-  { text: '4. Running dev server', link: '/tutorial/4-running-dev-server' },
-  { text: '5. Editing & saving a file', link: '/tutorial/5-editing-a-file-updating-the-iframe' },
-  { text: '6. Connect a terminal', link: '/tutorial/6-connect-a-terminal' },
-  { text: '7. Add interactivity', link: '/tutorial/7-add-interactivity' },
-];
-
-export const communityProjectsLinks = [
-  { text: 'All Projects', link: '/community-projects/all-projects' },
-  { text: 'Codeflow', link: '/community-projects/stackblitz-codeflow' },
-  { text: 're:tune', link: '/community-projects/retune' },
-  { text: 'schachnovelle', link: '/community-projects/schachnovelle' },
-  { text: 'SvelteKit', link: '/community-projects/sveltekit' },
-  { text: `Tech Educator's Toolkit`, link: '/community-projects/tech-educators-toolkit' },
-  { text: 'Web Publisher', link: '/community-projects/web-publisher' },
-];
-
-export const homeExternalLinks = [
-  {
-    title: 'Community',
-    description:
-      'Join our supportive community on Discord, ask questions, and share your StackBlitz projects.',
-    url: 'https://discord.com/invite/EQ7uJQxC',
-    large: true,
-    bgImgLight: '/img/theme/link-bg-squares-light.png',
-    bgImgDark: '/img/theme/link-bg-squares-dark.png',
-  },
-];
-
 export interface WCUsedBy {
   name: string;
   imgDark: string;
@@ -154,22 +110,6 @@ export const projectsUsingWCCardData: ProjectsUsingWCCardData[] = [
     },
   },
 ];
-
-const PEOPLE = {
-  DAN_JUTAN: 'DAN_JUTAN',
-  ERIC_SIMONS: 'ERIC_SIMONS',
-  RICH_HARRIS: 'RICH_HARRIS',
-  SYLWIA_VARGAS: 'SYLWIA_VARGAS',
-} as const;
-export type PeopleKeys = keyof typeof PEOPLE;
-
-export interface Person {
-  name: string;
-  href: string;
-  hrefAlias?: string;
-  imgPath?: string;
-  title?: string;
-}
 
 export interface TestimonialCardData {
   author: {
@@ -296,150 +236,5 @@ export const testimonialCardData: TestimonialCardData[] = [
       author: '/img/testimonials/manus_nijhoff.png',
     },
     quote: `Running chess in a terminal, running a terminal in the browser, check mate!<br/><strong>The best position to be in is a creative one and the StackBlitz WebContainers allow that.</strong>`,
-  },
-];
-
-export type CardLinkType = 'article'|'video';
-
-export type CommunityProjectCategory = 'ai'|'game'|'ide'|'lowCode'|'tutorial';
-export type CommunityProjectCategories = { [project in CommunityProjectCategory]: {
-  rgb: string;
-  title: string;
-  titlePlural?: string;
-}};
-
-export const communityProjectCategories: CommunityProjectCategories = {
-  ai: {
-    rgb: '255, 140, 0',
-    title: 'AI',
-  },
-  game: {
-    rgb: '188, 94, 255',
-    title: 'Game',
-    titlePlural: 'Games',
-  },
-  ide: {
-    rgb: '31, 143, 255',
-    title: 'Browser IDE',
-    titlePlural: 'Browser IDEs',
-  },
-  lowCode: {
-    rgb: '255, 9, 255',
-    title: 'Low-Code',
-  },
-  tutorial: {
-    rgb: '0, 188, 185',
-    title: 'Tutorial',
-    titlePlural: 'Tutorials',
-  },
-};
-
-export interface CommunityProjectCardData {
-  category: CommunityProjectCategory;
-  description: string;
-  title: string;
-  itemUrl: string;
-  thumbnailUrl: string;
-}
-
-export const communityProjectCardData: CommunityProjectCardData[] = [
-  {
-    category: 'tutorial',
-    title: 'SvelteKit',
-    description: 'A full educational experience of learning Svelte in the browser.',
-    itemUrl: '/community-projects/sveltekit',
-    thumbnailUrl: '/img/svelte-screen-light.png',
-  },
-  {
-    category: 'ide',
-    title: 'Codeflow',
-    description: 'Codeflow IDE</a> is a full-featured version of the desktop Visual Studio Code IDE supporting git commands, desktop extensions and a Node.js development server with terminal.',
-    itemUrl: '/community-projects/stackblitz-codeflow',
-    thumbnailUrl: '/img/community/codeflow.png',
-  },
-  {
-    category: 'ai',
-    title: 're:tune',
-    description: `The missing frontend for GPT-3, on a mission to empower everyone to build AI-first software at the speed of thought.`,
-    itemUrl: '/community-projects/retune',
-    thumbnailUrl: '/img/community/re_tune.png',
-  },
-  {
-    category: 'game',
-    title: 'schachnovelle',
-    description: `A command line chess game written in Node.js.`,
-    itemUrl: '/community-projects/schachnovelle',
-    thumbnailUrl: '/img/community/chess.png',
-  },
-  {
-    category: 'lowCode',
-    title: 'Web Publisher',
-    description: `A publishing tool that makes editing docs or blogs pleasant thanks to a realtime preview of the changes.`,
-    itemUrl: '/community-projects/web-publisher',
-    thumbnailUrl: '/img/community/web_publisher.png',
-  },
-  {
-    category: 'tutorial',
-    title: `Tech Educator's Toolkit`,
-    description: `Tech Educator's Toolkit is an early exploration of WebContainer API and of a new way to write interactive tutorials.`,
-    itemUrl: '/community-projects/tech-educators-toolkit',
-    thumbnailUrl: '/img/community/dan_jutan_talk.png',
-  }
-];
-
-export const footerSections = [
-  {
-    title: 'Workspaces',
-    items: [
-      { text: 'Popular', link: 'https://stackblitz.com/?starters=popular' },
-      { text: 'Frontend', link: 'https://stackblitz.com/?starters=frontend' },
-      { text: 'Backend', link: 'https://stackblitz.com/?starters=backend' },
-      { text: 'Fullstack', link: 'https://stackblitz.com/?starters=fullstack' },
-      { text: 'Vite', link: 'https://stackblitz.com/?starters=vite' },
-      { text: 'Docs, Blogs & Slides', link: 'https://stackblitz.com/?starters=docs' },
-      { text: 'Vanilla', link: 'https://stackblitz.com/?starters=vanilla' },
-    ],
-  },
-  {
-    title: 'Start a new Project',
-    items: [
-      { text: 'From a Workspace', link: 'https://stackblitz.com/?starters=popular' },
-      {
-        text: 'From a GitHub Repo',
-        link: 'https://developer.stackblitz.com/guides/user-guide/importing-projects#import-from-github',
-      },
-      {
-        text: 'From your computer',
-        link: 'https://developer.stackblitz.com/guides/user-guide/importing-projects#upload-from-your-computer',
-      },
-    ],
-  },
-  {
-    title: 'Product',
-    items: [
-      { text: 'Docs', link: 'https://developer.stackblitz.com/' },
-      { text: 'Enterprise', link: 'https://stackblitz.com/enterprise' },
-      { text: 'Pricing', link: 'https://stackblitz.com/membership' },
-      { text: 'Case Studies', link: 'https://stackblitz.com/case-studies/google' },
-    ],
-  },
-  {
-    title: 'Company',
-    items: [
-      { text: 'Blog', link: 'https://blog.stackblitz.com/' },
-      { text: 'Careers', link: 'https://stackblitz.com/careers' },
-      { text: 'Community', link: 'https://discord.gg/stackblitz' },
-      { text: 'Enterprise Sales', link: 'https://stackblitz.com/enterprise-contact' },
-      { text: 'Privacy', link: 'https://stackblitz.com/privacy-policy' },
-      { text: 'Terms of Service', link: 'https://stackblitz.com/terms-of-service' },
-    ],
-  },
-  {
-    title: 'Connect',
-    items: [
-      { text: 'GitHub', link: 'https://github.com/stackblitz/' },
-      { text: 'Twitter', link: 'https://twitter.com/stackblitz' },
-      { text: 'Discord', link: 'https://discord.gg/stackblitz' },
-    ],
   },
 ];
