+++ docs/.vitepress/theme/data/links.ts
@@ -1,0 +1,87 @@
+export const guideLinks = [
+  { text: 'Introduction', link: '/guides/introduction' },
+  { text: 'Quickstart', link: '/guides/quickstart' },
+  { text: 'Working with the File System', link: '/guides/working-with-the-file-system' },
+  { text: 'Running Processes', link: '/guides/running-processes' },
+  { text: 'Configuring Headers', link: '/guides/configuring-headers' },
+  { text: 'Troubleshooting', link: '/guides/troubleshooting'},
+  { text: 'Browser Support', link: '/guides/browser-support' },
+  { text: 'Browser Configuration', link: '/guides/browser-config' },
+];
+
+export const tutorialLinks = [
+  { text: '1. Build your first WebContainer app', link: '/tutorial/1-build-your-first-webcontainer-app' },
+  { text: '2. Setting up WebContainers', link: '/tutorial/2-setting-up-webcontainers' },
+  { text: '3. Installing dependencies', link: '/tutorial/3-installing-dependencies' },
+  { text: '4. Running dev server', link: '/tutorial/4-running-dev-server' },
+  { text: '5. Editing & saving a file', link: '/tutorial/5-editing-a-file-updating-the-iframe' },
+  { text: '6. Connect a terminal', link: '/tutorial/6-connect-a-terminal' },
+  { text: '7. Add interactivity', link: '/tutorial/7-add-interactivity' },
+];
+
+export const communityProjectsLinks = [
+  { text: 'All Projects', link: '/community-projects/all-projects' },
+  { text: 're:tune', link: '/community-projects/retune' },
+  { text: 'schachnovelle', link: '/community-projects/schachnovelle' },
+  { text: 'StackBlitz Codeflow', link: '/community-projects/stackblitz-codeflow' },
+  { text: 'StackBlitz Web Publisher', link: '/community-projects/web-publisher' },
+  { text: 'SvelteKit', link: '/community-projects/sveltekit' },
+  { text: `Tech Educator's Toolkit`, link: '/community-projects/tech-educators-toolkit' },
+];
+
+export const footerSections = [
+  {
+    title: 'Workspaces',
+    items: [
+      { text: 'Popular', link: 'https://stackblitz.com/?starters=popular' },
+      { text: 'Frontend', link: 'https://stackblitz.com/?starters=frontend' },
+      { text: 'Backend', link: 'https://stackblitz.com/?starters=backend' },
+      { text: 'Fullstack', link: 'https://stackblitz.com/?starters=fullstack' },
+      { text: 'Vite', link: 'https://stackblitz.com/?starters=vite' },
+      { text: 'Docs, Blogs & Slides', link: 'https://stackblitz.com/?starters=docs' },
+      { text: 'Vanilla', link: 'https://stackblitz.com/?starters=vanilla' },
+    ],
+  },
+  {
+    title: 'Start a new Project',
+    items: [
+      { text: 'From a Workspace', link: 'https://stackblitz.com/?starters=popular' },
+      {
+        text: 'From a GitHub Repo',
+        link: 'https://developer.stackblitz.com/guides/user-guide/importing-projects#import-from-github',
+      },
+      {
+        text: 'From your computer',
+        link: 'https://developer.stackblitz.com/guides/user-guide/importing-projects#upload-from-your-computer',
+      },
+    ],
+  },
+  {
+    title: 'Product',
+    items: [
+      { text: 'Docs', link: 'https://developer.stackblitz.com/' },
+      { text: 'Enterprise', link: 'https://stackblitz.com/enterprise' },
+      { text: 'Pricing', link: 'https://stackblitz.com/membership' },
+      { text: 'Case Studies', link: 'https://stackblitz.com/case-studies/google' },
+    ],
+  },
+  {
+    title: 'Company',
+    items: [
+      { text: 'Blog', link: 'https://blog.stackblitz.com/' },
+      { text: 'Careers', link: 'https://stackblitz.com/careers' },
+      { text: 'Community', link: 'https://discord.gg/stackblitz' },
+      { text: 'Enterprise Sales', link: 'https://stackblitz.com/enterprise-contact' },
+      { text: 'Privacy', link: 'https://stackblitz.com/privacy-policy' },
+      { text: 'Terms of Service', link: 'https://stackblitz.com/terms-of-service' },
+    ],
+  },
+  {
+    title: 'Connect',
+    items: [
+      { text: 'GitHub', link: 'https://github.com/stackblitz/' },
+      { text: 'Twitter', link: 'https://twitter.com/stackblitz' },
+      { text: 'Discord', link: 'https://discord.gg/stackblitz' },
+    ],
+  },
+];
